% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_NCBI_functions.R
\name{get_citations}
\alias{get_citations}
\title{Get a citations available from XML}
\usage{
get_citations(xml)
}
\arguments{
\item{xml}{a character with XML code}
}
\value{
a list with all the PMID for all the articles cited in the XML if it's possible. Otherwise it's NULL
}
\description{
Get a citations available from XML
}
