% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_NCBI_functions.R
\name{get_from_xml}
\alias{get_from_xml}
\title{Get a variable from XML
variables availables are "title", "authors", "year", "journal", "volume", "issue", "pages", "key_words", "doi", "pmid", "abstract"}
\usage{
get_from_xml(xml, what = "title")
}
\arguments{
\item{xml}{a character with XML code}

\item{what}{the name of the variable you want}
}
\value{
a list with all the "variable selected" for all the articles in the XML
}
\description{
Get a variable from XML
variables availables are "title", "authors", "year", "journal", "volume", "issue", "pages", "key_words", "doi", "pmid", "abstract"
}
\examples{
library(magrittr)
gets_ids("oyster herpesvirus") \%>\%
  get_xml() \%>\%
  get_from_xml("title")
}
