% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_NCBI_functions.R
\name{get_ids}
\alias{get_ids}
\title{Get ID (or PMID) from a query to NCBI}
\usage{
get_ids(query, retmax = 1000)
}
\arguments{
\item{query}{a query to enter in NCBI}

\item{retmax}{the maximum number of results (default = 1000)}
}
\value{
a vector with ids
}
\description{
Get ID (or PMID) from a query to NCBI
}
\examples{
gets_ids("oyster herpesvirus")
}
