% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_NCBI_functions.R
\name{get_xml}
\alias{get_xml}
\title{Get XML code from NCBI giving a ids vector}
\usage{
get_xml(ids)
}
\arguments{
\item{ids}{a vector with articles ids (PMID)}
}
\value{
a character with XML code
}
\description{
Get XML code from NCBI giving a ids vector
}
\examples{
gets_ids("oyster herpesvirus") \%>\%
  get_xml()
}
