% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_df}
\alias{index_df}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 399971 rows and 24 variables:
\describe{
\item{Rank}{SJR Rank of the journal}
\item{Sourceid}{id from SJR}
\item{journal}{Complete name of the journal}
\item{Type}{only journal are kept here, so, useless columns}
\item{SJR}{SJR index}
\item{Issn}{Issn number}
\item{Issn_2}{Issn_2 number}
\item{SJR Best Quartile}{SJR best quartile}
\item{H index}{H index of the journal}
\item{Total Docs. (3years)}{Total number of documents for 3 years}
\item{Total Refs.}{Total number of references}
\item{Total Cites (3years)}{Total cites number for 3 years}
\item{Citable Docs. (3years)}{Number of citable documents fr 3 years}
\item{Cites / Doc. (2years)}{Number of cites by document for 2 years}
\item{Ref. / Doc.}{Number of references by document}
\item{Country}{Country}
\item{Publisher}{Publisher name}
\item{Coverage}{Coverage time of the journal (since it has been created)}
\item{Categories}{Categories, field of the journal}
\item{year}{year}
\item{Total_Docs}{Total_Docs}
\item{Total Cites}{Total Cites}
\item{Journal Impact Factor}{Journal Impact Factor index}
\item{Eigenfactor Score}{Eigenfactor Score index}
}}
\source{
\url{https://www.scimagojr.com/journalrank.php}
}
\usage{
index_df
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
