% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_NCBI_functions.R
\name{make_df}
\alias{make_df}
\title{Make a data frame with ID and an other variable of interest ("title", "authors", "year", "journal", "volume", "issue", "pages", "key_words", "doi", "pmid", "abstract")}
\usage{
make_df(query, xml, var)
}
\arguments{
\item{query}{a query to enter to NCBI}

\item{xml}{a character with XML code}

\item{var}{the name of the variable you want}
}
\value{
a tibble with 2 columns, id and "variable selected"
}
\description{
Make a data frame with ID and an other variable of interest ("title", "authors", "year", "journal", "volume", "issue", "pages", "key_words", "doi", "pmid", "abstract")
}
\examples{
query <- "oyster herpesvirus"
xml <- get_ids(query) \%>\%
         get_xml()
make_df(query, xml, "abstract" )
}
